remove osinfo and other legacy
