+++ conan/tools/apple/xcodedeps.py
@@ -3,7 +3,6 @@ import textwrap
 
 from jinja2 import Template
 
-from conan.tools._check_build_profile import check_using_build_profile
 from conans.errors import ConanException
 from conans.util.files import load, save
 from conan.tools.apple.apple import to_apple_arch
@@ -77,7 +76,6 @@ class XcodeDeps(object):
         # TODO: check if it makes sense to add a subsetting for sdk version
         #  related to: https://github.com/conan-io/conan/issues/9608
         self.os_version = conanfile.settings.get_safe("os.version")
-        check_using_build_profile(self._conanfile)
 
     def generate(self):
         if self.configuration is None:
