+++ conan/tools/cmake/cmake.py
@@ -1,12 +1,12 @@
 import os
 import platform
 
+from conan.tools import args_to_string
 from conan.tools.build import build_jobs
 from conan.tools.cmake.utils import is_multi_configuration
 from conan.tools.files import load_toolchain_args, chdir, mkdir
 
 from conan.tools.microsoft.msbuild import msbuild_verbosity_cmd_line_arg
-from conans.client.tools.oss import args_to_string
 from conans.errors import ConanException
