+++ conan/tools/google/bazeldeps.py
@@ -2,14 +2,12 @@ import textwrap
 
 from jinja2 import Template
 
-from conan.tools._check_build_profile import check_using_build_profile
 from conans.util.files import save
 
 
 class BazelDeps(object):
     def __init__(self, conanfile):
         self._conanfile = conanfile
-        check_using_build_profile(self._conanfile)
 
     def generate(self):
         local_repositories = []
