+++ conan/tools/google/toolchain.py
@@ -1,4 +1,3 @@
-from conan.tools._check_build_profile import check_using_build_profile
 from conan.tools.files import save_toolchain_args
 
 
@@ -7,7 +6,6 @@ class BazelToolchain(object):
     def __init__(self, conanfile, namespace=None):
         self._conanfile = conanfile
         self._namespace = namespace
-        check_using_build_profile(self._conanfile)
 
     def generate(self):
         save_toolchain_args({
