+++ conan/tools/meson/toolchain.py
@@ -2,7 +2,6 @@ import textwrap
 
 from jinja2 import Template
 
-from conan.tools._check_build_profile import check_using_build_profile
 from conan.tools.env import VirtualBuildEnv
 from conan.tools.microsoft import VCVars
 from conans.client.build.cppstd_flags import cppstd_from_settings
@@ -117,8 +116,6 @@ class MesonToolchain(object):
                                              self._env_array('CXXFLAGS'))
         self.cpp_link_args = self._to_meson_value(self._env_array('LDFLAGS'))
         self.pkg_config_path = "'%s'" % self._conanfile.generators_folder
-
-        check_using_build_profile(self._conanfile)
 
     def _get_backend(self, recipe_backend):
         # Returns the name of the backend used by Meson
