--- conans/test/unittests/client/tools/os_info/__init__.py
@@ -1,3 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-# vim: tabstop=8 expandtab shiftwidth=4 softtabstop=4
